+++ dataconf/__init__.py
@@ -4,12 +4,14 @@ from dataconf.main import dump
 from dataconf.main import dumps
 from dataconf.main import env
 from dataconf.main import file
+from dataconf.main import HOCON
 from dataconf.main import load
 from dataconf.main import loads
 from dataconf.main import multi
 from dataconf.main import parse
 from dataconf.main import string
 from dataconf.main import url
+from dataconf.main import YAML
 from dataconf.version import __version__
 
 __all__ = [
@@ -25,5 +27,7 @@ __all__ = [
     "cli",
     "multi",
     "parse",
+    "YAML",
+    "HOCON",
     "__version__",
 ]