+++ dataconf/main.py
@@ -1,7 +1,12 @@
 import inspect
+import contextlib
 import os
+from typing import Any, Optional
+from typing import Dict
 from typing import List
 from typing import Type
+from urllib.parse import urlparse
+from urllib.request import urlopen
 
 from dataconf import utils
 from dataconf.exceptions import MalformedConfigException
@@ -9,6 +14,10 @@ from pyhocon import ConfigFactory
 from pyhocon import HOCONConverter
 from pyhocon.config_parser import ConfigTree
 import pyparsing
+from yaml import safe_load
+
+HOCON = 1
+YAML = 2
 
 
 def inject_callee_scope(func):
@@ -41,7 +50,6 @@ def parse(
     globalns=None,
     localns=None,
 ):
-
     try:
         return utils.__parse(
             conf, clazz, "", strict, ignore_unexpected, globalns, localns
@@ -66,23 +74,39 @@ class Multi:
         self.strict = strict
 
     def env(self, prefix: str) -> "Multi":
-        self.strict = False
         data = env_vars_parse(prefix, os.environ)
-        return self.dict(data)
+        return Multi(self.confs, strict=False).dict(data)
 
     def dict(self, obj: str) -> "Multi":
         conf = ConfigFactory.from_dict(obj)
         return Multi(self.confs + [conf], self.strict)
 
-    def string(self, s: str) -> "Multi":
+    def string(self, s: str, loader: str = HOCON) -> "Multi":
+        if loader == YAML:
+            data = safe_load(s)
+            return self.dict(data)
+
         conf = ConfigFactory.parse_string(s)
         return Multi(self.confs + [conf], self.strict)
 
-    def url(self, uri: str) -> "Multi":
-        conf = ConfigFactory.parse_URL(uri)
+    def url(self, uri: str, timeout: int = 10) -> "Multi":
+        path = urlparse(uri).path
+        if path.endswith(".yaml") or path.endswith(".yml"):
+            with contextlib.closing(urlopen(uri, timeout=timeout)) as fd:
+                s = fd.read().decode("utf-8")
+            return self.string(s, loader=YAML)
+
+        conf = ConfigFactory.parse_URL(uri, timeout=timeout, required=True)
         return Multi(self.confs + [conf], self.strict)
 
-    def file(self, path: str) -> "Multi":
+    def file(self, path: str, loader: Optional[str] = None) -> "Multi":
+        if loader == YAML or (
+            loader is None and (path.endswith(".yaml") or path.endswith(".yml"))
+        ):
+            with open(path, "r") as f:
+                data = safe_load(f)
+            return self.dict(data)
+
         conf = ConfigFactory.parse_file(path)
         return Multi(self.confs + [conf], self.strict)
 
@@ -107,13 +131,13 @@ def env(prefix: str, clazz: Type, **kwargs):
 
 
 @inject_callee_scope
-def dict(obj: str, clazz: Type, **kwargs):
+def dict(obj: Dict[str, Any], clazz: Type, **kwargs):
     return multi.dict(obj).on(clazz, **kwargs)
 
 
 @inject_callee_scope
-def string(s: str, clazz: Type, **kwargs):
-    return multi.string(s).on(clazz, **kwargs)
+def string(s: str, clazz: Type, loader: str = HOCON, **kwargs):
+    return multi.string(s, loader).on(clazz, **kwargs)
 
 
 @inject_callee_scope
@@ -122,8 +146,8 @@ def url(uri: str, clazz: Type, **kwargs):
 
 
 @inject_callee_scope
-def file(path: str, clazz: Type, **kwargs):
-    return multi.file(path).on(clazz, **kwargs)
+def file(path: str, clazz: Type, loader: Optional[str] = None, **kwargs):
+    return multi.file(path, loader).on(clazz, **kwargs)
 
 
 @inject_callee_scope
@@ -132,13 +156,13 @@ def cli(argv: List[str], clazz: Type, **kwargs):
 
 
 @inject_callee_scope
-def load(path: str, clazz: Type, **kwargs):
-    return file(path, clazz, **kwargs)
+def load(path: str, clazz: Type, loader: Optional[str] = None, **kwargs):
+    return file(path, clazz, loader, **kwargs)
 
 
 @inject_callee_scope
-def loads(s: str, clazz: Type, **kwargs):
-    return string(s, clazz, **kwargs)
+def loads(s: str, clazz: Type, loader: str = HOCON, **kwargs):
+    return string(s, clazz, loader, **kwargs)
 
 
 def dump(file: str, instance: object, out: str):
@@ -147,7 +171,7 @@ def dump(file: str, instance: object, out: str):
 
 
 def dumps(instance: object, out: str):
-    conf = utils.generate(instance, "")
+    conf = utils.__generate(instance, "")
 
     if out:
         if out.lower() == "hocon":